@@ -1,53 +1,49 @@
 {
-  "name": "app",
-  "version": "1.0.0",
+  "name": "quasar-vite-update",
+  "version": "0.0.1",
   "description": "A Quasar Project",
   "productName": "Quasar App",
-  "author": "",
+  "author": "dev-uros <uros.minic@kentkart.rs>",
+  "type": "module",
   "private": true,
   "scripts": {
-    "lint": "eslint --ext .js,.ts,.vue ./",
+    "lint": "eslint -c ./eslint.config.js \"./src*/**/*.{ts,js,cjs,mjs,vue}\"",
     "format": "prettier --write \"**/*.{js,ts,vue,scss,html,md,json}\" --ignore-path .gitignore",
-    "test": "vitest --run --reporter verbose",
-    "test:watch": "vitest",
-    "page": "quasar new p -ts-composition-setup",
-    "component": "quasar new c -ts-composition-setup",
-    "store": "quasar new s",
-    "build": "quasar clean && quasar build",
-    "postbuild": "node src/notifyClientsNewAppVersionAvailable.js",
-    "qmodule": "qmodule"
+    "test": "echo \"No test specified\" && exit 0",
+    "dev": "quasar dev",
+    "build": "quasar build",
+    "postinstall": "quasar prepare"
   },
   "dependencies": {
-    "@quasar/extras": "^1.16.12",
-    "@vueuse/components": "^10.11.0",
-    "@vueuse/core": "^10.11.0",
-    "axios": "^1.7.2",
-    "dotenv": "^16.4.5",
-    "moment-timezone": "^0.5.45",
-    "nanoid": "^5.0.7",
+    "@quasar/extras": "^1.16.4",
+    "axios": "^1.2.1",
+    "dotenv": "^16.4.7",
     "pinia": "^2.0.11",
-    "quasar": "^2.16.6",
-    "vue": "^3.4.35",
-    "vue-json-pretty": "^2.4.0",
-    "vue-router": "^4.4.0"
-  },
-  "devDependencies": {
+    "quasar": "^2.16.0",
     "quasar-easy-modules": "^1.0.17",
     "quasar-easy-validate": "^1.0.2",
-    "@quasar/app-vite": "^1.9.4",
-    "@types/node": "^22.1.0",
-    "@typescript-eslint/eslint-plugin": "^5.10.0",
-    "@typescript-eslint/parser": "^5.10.0",
+    "vue": "^3.4.18",
+    "vue-router": "^4.0.12"
+  },
+  "devDependencies": {
+    "@eslint/js": "^9.14.0",
+    "@quasar/app-vite": "^2.0.0",
+    "@types/node": "^20.5.9",
+    "@vue/devtools": "^7.6.8",
+    "@vue/eslint-config-prettier": "^10.1.0",
+    "@vue/eslint-config-typescript": "^14.1.3",
     "autoprefixer": "^10.4.2",
-    "eslint": "8.22.0",
-    "eslint-config-prettier": "^8.1.0",
-    "eslint-plugin-vue": "^9.0.0",
-    "prettier": "^2.5.1",
-    "typescript": "^5.5.4",
-    "vitest": "^2.0.4"
+    "eslint": "^9.14.0",
+    "eslint-plugin-vue": "^9.30.0",
+    "globals": "^15.12.0",
+    "prettier": "^3.3.3",
+    "typescript": "~5.5.3",
+    "vite-plugin-checker": "^0.8.0",
+    "vite-plugin-vue-devtools": "^7.6.8",
+    "vue-tsc": "^2.0.29"
   },
   "engines": {
-    "node": "^18 || ^16 || ^14.19 || ^20",
+    "node": "^28 || ^26 || ^24 || ^22 || ^20 || ^18",
     "npm": ">= 6.13.4",
     "yarn": ">= 1.21.1"
   }
