@@ -1,8 +1,8 @@
 import {HttpMethod, useFetch} from "src/modules/shared/utils/fetch";
-import {StoreNewUserResponse} from "../types";
+import type {StoreNewUserResponse} from "../types";
 
 export async function setAccountState(userId: number){
-  const {message, data: {user}} = await useFetch<StoreNewUserResponse>({
+  const {data: {user}} = await useFetch<StoreNewUserResponse>({
     url: `/administration/users/set-status/${userId}`,
     method: HttpMethod.PATCH
   })
