@@ -1,5 +1,5 @@
 import {HttpMethod, useFetch} from "src/modules/shared/utils/fetch";
-import {StoreNewUserRequest, StoreNewUserResponse} from "../types";
+import type {StoreNewUserRequest, StoreNewUserResponse} from "../types";
 
 export async function storeUser(request: StoreNewUserRequest) {
   const {data: {user}} = await useFetch<StoreNewUserResponse, StoreNewUserRequest>({
