@@ -3,24 +3,19 @@
     <q-card class="my-card">
       <q-card-section class="bg-primary">
         <q-img src="/images/blank-image.jpg" style="max-height: 150px"></q-img>
-        <q-separator color="white"/>
+        <q-separator color="white" />
       </q-card-section>
       <q-card-section>
-        <div class="text-h6">
-          <q-avatar icon="account_circle" /> Aktivacija naloga
-        </div>
+        <div class="text-h6"><q-avatar icon="account_circle" /> Aktivacija naloga</div>
       </q-card-section>
       <q-card-section>
-        <q-form
-          @submit="onSubmit"
-          class="q-gutter-md"
-        >
+        <q-form @submit="onSubmit" class="q-gutter-md">
           <q-input
             filled
-            v-model.trim='activateAccountForm.password'
-            label='Lozinka'
-            hint='Unesite lozinku'
-            color='black'
+            v-model.trim="activateAccountForm.password"
+            label="Lozinka"
+            hint="Unesite lozinku"
+            color="black"
             :type="isPwd ? 'password' : 'text'"
             :rules="passwordRules"
             lazy-rules
@@ -29,65 +24,77 @@
             <template v-slot:append>
               <q-icon
                 :name="isPwd ? 'visibility_off' : 'visibility'"
-                class='cursor-pointer'
-                @click='isPwd = !isPwd'
+                class="cursor-pointer"
+                @click="isPwd = !isPwd"
               />
             </template>
           </q-input>
 
-          <q-list dense >
+          <q-list dense>
             <q-item>
               <q-item-section avatar>
-                <q-icon :color="upperCasePassed ? 'positive' : 'negative'" :name="upperCasePassed ? 'done' : 'close'" />
+                <q-icon
+                  :color="upperCasePassed ? 'positive' : 'negative'"
+                  :name="upperCasePassed ? 'done' : 'close'"
+                />
               </q-item-section>
 
               <q-item-section>Veliko slovo</q-item-section>
             </q-item>
 
-            <q-item >
+            <q-item>
               <q-item-section avatar>
-                <q-icon :color="symbolPassed ? 'positive' : 'negative'" :name="symbolPassed ? 'done' : 'close'" />
+                <q-icon
+                  :color="symbolPassed ? 'positive' : 'negative'"
+                  :name="symbolPassed ? 'done' : 'close'"
+                />
               </q-item-section>
 
               <q-item-section>Simbol</q-item-section>
             </q-item>
 
-            <q-item >
+            <q-item>
               <q-item-section avatar>
-                <q-icon :color="minCharactersPassed ? 'positive' : 'negative'" :name="minCharactersPassed ? 'done' : 'close'" />
+                <q-icon
+                  :color="minCharactersPassed ? 'positive' : 'negative'"
+                  :name="minCharactersPassed ? 'done' : 'close'"
+                />
               </q-item-section>
 
               <q-item-section>Barem osam karaktera (!,@,#,$,%,^,&,*)</q-item-section>
             </q-item>
 
-            <q-item >
+            <q-item>
               <q-item-section avatar>
-                <q-icon :color="numberPassed ? 'positive' : 'negative'" :name="numberPassed ? 'done' : 'close'" />
+                <q-icon
+                  :color="numberPassed ? 'positive' : 'negative'"
+                  :name="numberPassed ? 'done' : 'close'"
+                />
               </q-item-section>
               <q-item-section>Barem jedan broj</q-item-section>
             </q-item>
           </q-list>
           <q-input
             filled
-            v-model.trim='activateAccountForm.password_confirmation'
-            label='Potvrda lozinke'
-            hint='Potvrdite unetu lozinku'
+            v-model.trim="activateAccountForm.password_confirmation"
+            label="Potvrda lozinke"
+            hint="Potvrdite unetu lozinku"
             lazy-rules
-            color='black'
+            color="black"
             :type="isPwdConfirm ? 'password' : 'text'"
             :rules="passwordConfirmationRules"
           >
             <template v-slot:append>
               <q-icon
                 :name="isPwdConfirm ? 'visibility_off' : 'visibility'"
-                class='cursor-pointer'
-                @click='isPwdConfirm = !isPwdConfirm'
+                class="cursor-pointer"
+                @click="isPwdConfirm = !isPwdConfirm"
               />
             </template>
           </q-input>
 
           <div>
-            <q-btn label="Aktiviraj nalog" type="submit" color="primary"/>
+            <q-btn label="Aktiviraj nalog" type="submit" color="primary" />
           </div>
         </q-form>
       </q-card-section>
@@ -96,77 +103,82 @@
 </template>
 
 <script setup lang="ts">
-import { useRoute, useRouter } from 'vue-router';
-import {reactive, ref} from 'vue';
-import {Validator} from "quasar-easy-validate";
-import {useAuthStore} from "../store";
+import { useRoute, useRouter } from 'vue-router'
+import { reactive, ref } from 'vue'
+import { Validator } from 'quasar-easy-validate'
+import { useAuthStore } from '../store'
 
-const route = useRoute();
-const router = useRouter();
-const userStore = useAuthStore();
+const route = useRoute()
+const router = useRouter()
+const userStore = useAuthStore()
 
-const isPwd = ref(true);
-const isPwdConfirm = ref(true);
+const isPwd = ref(true)
+const isPwdConfirm = ref(true)
 const activateAccountForm = reactive({
   password: '',
   password_confirmation: '',
   email: route.query.email!.toString(),
-  token: route.params.token as string
+  token: route.params.token as string,
 })
 
-const passwordRules = [(password: string) => (new Validator(password)
-  .required('Lozinka je obavezna')
-  .password('Lozinka mora da sadrži barem jedno veliko slovo, malo slovo, cifru i specijalan karakter (!,@,#,$,%,^,&,*).')
-  .validate())
-];
+const passwordRules = [
+  (password: string) =>
+    new Validator(password)
+      .required('Lozinka je obavezna')
+      .password(
+        'Lozinka mora da sadrži barem jedno veliko slovo, malo slovo, cifru i specijalan karakter (!,@,#,$,%,^,&,*).',
+      )
+      .validate(),
+]
 
 function containsUpperCase(password: string) {
-  return /[A-Z]/.test(password);
+  return /[A-Z]/.test(password)
 }
 
-function containsSymbol(password:string) {
+function containsSymbol(password: string) {
   // Specify the list of symbols you want to check for
-  const symbolList = /[\!\@\#\$\%\^\&\*]/;
+  const symbolList = /[!@#$%^&*]/
 
-  return symbolList.test(password);
+  return symbolList.test(password)
 }
 
-function hasNeededAmountOfCharacters(password: string, numberOfCharacters: number){
-  return password.length >= numberOfCharacters;
+function hasNeededAmountOfCharacters(password: string, numberOfCharacters: number) {
+  return password.length >= numberOfCharacters
 }
 
 function containsNumber(password: string) {
-  return /\d/.test(password);
+  return /\d/.test(password)
 }
 
-const upperCasePassed = ref(false);
-const symbolPassed = ref(false);
-const minCharactersPassed = ref(false);
-const numberPassed = ref(false);
+const upperCasePassed = ref(false)
+const symbolPassed = ref(false)
+const minCharactersPassed = ref(false)
+const numberPassed = ref(false)
 
 const passwordCheck = () => {
+  upperCasePassed.value = containsUpperCase(activateAccountForm.password)
 
-  upperCasePassed.value = containsUpperCase(activateAccountForm.password);
+  symbolPassed.value = containsSymbol(activateAccountForm.password)
 
-  symbolPassed.value = containsSymbol(activateAccountForm.password);
-
-  numberPassed.value = containsNumber(activateAccountForm.password);
+  numberPassed.value = containsNumber(activateAccountForm.password)
 
   minCharactersPassed.value = hasNeededAmountOfCharacters(activateAccountForm.password, 8)
-
 }
-const passwordConfirmationRules = [(passwordConfirmation: string) => (new Validator(passwordConfirmation)
-  .required('Morate potvrditi lozinku')
-  .password('Lozinka mora da sadrži barem jedno veliko slovo, malo slovo, cifru i specijalan karakter (!,@,#,$,%,^,&,*).')
-  .sameAs(activateAccountForm.password, 'Lozinke se ne podudaraju')
-  .validate())
-];
+const passwordConfirmationRules = [
+  (passwordConfirmation: string) =>
+    new Validator(passwordConfirmation)
+      .required('Morate potvrditi lozinku')
+      .password(
+        'Lozinka mora da sadrži barem jedno veliko slovo, malo slovo, cifru i specijalan karakter (!,@,#,$,%,^,&,*).',
+      )
+      .sameAs(activateAccountForm.password, 'Lozinke se ne podudaraju')
+      .validate(),
+]
 
-async function onSubmit(){
-
-  await userStore.activateAccount(activateAccountForm);
+async function onSubmit() {
+  await userStore.activateAccount(activateAccountForm)
   await router.replace({
-    name: 'login'
+    name: 'login',
   })
 }
 </script>
@@ -175,6 +187,6 @@ async function onSubmit(){
 .my-card {
   width: 100%;
   max-width: 500px;
-  margin: 0 auto
+  margin: 0 auto;
 }
 </style>
