@@ -3,26 +3,23 @@
     <q-card class="my-card">
       <q-card-section class="bg-primary">
         <q-img src="/images/blank-image.jpg" style="max-height: 150px"></q-img>
-        <q-separator color="white"/>
+        <q-separator color="white" />
       </q-card-section>
       <q-card-section>
-        <q-form
-          class="q-gutter-md"
-          @submit="onSubmit"
-        >
+        <q-form class="q-gutter-md" @submit="onSubmit">
           <q-input
             filled
             v-model.trim="loginForm.email"
-            hint='Unesite Vaš e-mail'
+            hint="Unesite Vaš e-mail"
             label="Email"
             lazy-rules
             :rules="emailRules"
           />
           <q-input
             filled
-            label='Lozinka'
-            hint='Unesite Vašu loziniku'
-            color='black'
+            label="Lozinka"
+            hint="Unesite Vašu loziniku"
+            color="black"
             :type="isPwd ? 'password' : 'text'"
             v-model.trim="loginForm.password"
             :rules="passwordRules"
@@ -30,62 +27,68 @@
             <template v-slot:append>
               <q-icon
                 :name="isPwd ? 'visibility_off' : 'visibility'"
-                class='cursor-pointer'
-                @click='isPwd = !isPwd'
+                class="cursor-pointer"
+                @click="isPwd = !isPwd"
               />
             </template>
           </q-input>
           <div class="button-container">
-            <q-btn label='Uloguj se' type='submit' color='primary'/>
-            <q-btn label='Zaboravljena lozinka?' type='button' color='primary' flat class='q-ml-sm'
-                   @click='forgottenPassword'/>
+            <q-btn label="Uloguj se" type="submit" color="primary" />
+            <q-btn
+              label="Zaboravljena lozinka?"
+              type="button"
+              color="primary"
+              flat
+              class="q-ml-sm"
+              @click="forgottenPassword"
+            />
           </div>
         </q-form>
       </q-card-section>
     </q-card>
-
   </q-page>
 </template>
 
 <script setup lang="ts">
-import {reactive, ref} from 'vue';
-import {useRouter} from 'vue-router';
-import {Validator} from "quasar-easy-validate";
-import {useAuthStore} from "../store";
+import { reactive, ref } from 'vue'
+import { useRouter } from 'vue-router'
+import { Validator } from 'quasar-easy-validate'
+import { useAuthStore } from '../store'
 
-const router = useRouter();
-const userStore = useAuthStore();
+const router = useRouter()
+const userStore = useAuthStore()
 
 const loginForm = reactive({
   email: '',
-  password: ''
-});
+  password: '',
+})
 
-const emailRules = [(email: string) => (new Validator(email)
-  .required('E-mail je obavezan')
-  .stringMin(3, 'E-mail može imati minimalno 3 karaktera')
-  .stringMax(255, 'E-mail može imati maksimalno 255 karaktera')
-  .email('Email nije validan')
-  .validate())
-];
-
-const isPwd = ref(true);
-
-const passwordRules = [(password: string) => (new Validator(password)
-  .required('Lozinka je obavezna')
-  .validate())
+const emailRules = [
+  (email: string) =>
+    new Validator(email)
+      .required('E-mail je obavezan')
+      .stringMin(3, 'E-mail može imati minimalno 3 karaktera')
+      .stringMax(255, 'E-mail može imati maksimalno 255 karaktera')
+      .email('Email nije validan')
+      .validate(),
 ]
 
-async function onSubmit(){
-  await userStore.login(loginForm);
+const isPwd = ref(true)
+
+const passwordRules = [
+  (password: string) => new Validator(password).required('Lozinka je obavezna').validate(),
+]
+
+async function onSubmit() {
+  await userStore.login(loginForm)
   await router.replace({
-    name: 'home'
+    name: 'home',
   })
 }
 
-function forgottenPassword(){
+function forgottenPassword() {
   router.push({
-    name: 'forgot-password'
+    name: 'forgot-password',
   })
 }
 </script>
@@ -96,7 +99,7 @@ function forgottenPassword(){
   max-width: 500px;
   margin: 0 auto;
 }
-.button-container{
+.button-container {
   display: flex;
   justify-content: space-between;
 }
