@@ -1,5 +1,4 @@
 <template>
   <h3>Hello From Component</h3>
 </template>
-<script setup lang="ts">
-</script>
+<script setup lang="ts"></script>
