@@ -1,5 +1,5 @@
 
-import {RouteRecordRaw} from "vue-router";
+import type {RouteRecordRaw} from "vue-router";
 
 import auth from "../shared/middleware/auth";
 
