@@ -1,7 +1,7 @@
 <template>
   <q-item>
     <q-item-section avatar>
-      <q-icon :name="icon" size="30px"/>
+      <q-icon :name="icon" size="30px" />
     </q-item-section>
     <q-item-section>
       <q-item-label class="text-h5">{{ title }}</q-item-label>
@@ -12,15 +12,13 @@
 
 <script setup lang="ts">
 interface BaseHeaderPropsInterface {
-  icon: string,
+  icon: string
   title: string
 }
 withDefaults(defineProps<BaseHeaderPropsInterface>(), {
   icon: 'edit_note',
-  title: 'Basic Header Title'
+  title: 'Basic Header Title',
 })
 </script>
 
-<style scoped>
-
-</style>
+<style scoped></style>
